+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # CHANGELOG
 
+## 1.8.4 | 00.00.2020 <gulp, js, mautic, anchor & inline fs>
+* add gulp-babel to allow const and let
+* add use strict to javascript
+* add body and frontend class to additional JS css
+* check element existens before setting event listener
+* mautic class added to configuration selection
+* IE javascript file for fixes
+* Anchor position for sticky header // css solution
+* Inline font size responsive solution
+
 ## 1.7.4 | 11.09.2020 <ajax safty, javascript 1.0 release>
 * fix video max width
 * ajax safty